requirements.txt
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
greenlet==3.2.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
outcome==1.3.0.post0
pandas==2.2.3
pydantic==2.11.3
pydantic_core==2.33.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
sqlmodel==0.0.24
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
